// src/buttons/dropdown/types.ts
import { ReactNode } from 'react';

interface IDropdownOptions {
  value: number | string;
  label: string;
  action: (value: any) => void;
  component?: ReactNode;
}

export interface IDropdownProps {
  children?: ReactNode;
  disabled?: boolean;
  options?: IDropdownOptions[];
  dropdown?: boolean;
  variant?: 'button' | 'sorting';
  width?: string;
  optionswidth?: string;
  optionsvalue?: number | string;
  value?: number | string;
  showOption?: boolean;
}